* libdb/db_btree.c (btree_findkey): Cast key.dsize to int when using it as an fprintf() field width.

+++ libdb/db_btree.c
@@ -219,7 +219,7 @@ static __inline__ datum btree_findkey(DB *dbf, u_int flags)
 		if (debug)
 			fprintf (stderr, "Corrupt database! Already seen %*s. "
 					 "Attempting to recover ...\n",
-				 key.dsize, key.dptr);
+				 (int) key.dsize, key.dptr);
 		key.dptr = NULL;
 		key.dsize = 0;
 		return key;
